Fix typo in no-unsupported-features VERSION_MAP (0.1 -> 0.10)
The rule no-unsupported-features allows to configure a specific version
through the rule options:
https://github.com/mysticatea/eslint-plugin-node/blob/master/docs/rules/no-unsupported-features.md#version

As per the documentation above it should allow to configure the version
to `0.10` to allow only features in node v0.10, but this would have no
effect, because the only available key for v0.10 was `0.1`.

This commit changes the key of the VERSION_MAP from the incorrect value
of `0.1` to `0.10`.