use is-core-module to support node: scheme
As requested in #275, recognize Node.js core modules referenced using
the [`node:`](https://nodejs.org/api/esm.html#esm_node_imports) URL
scheme.

Since `resolve.isCore` was moved to `is-core-module` in
https://github.com/browserify/resolve/commit/7c26483576f7a44573ddc61cd5d51519deb674c5
and `is-core-module@2.3.0` added support for `node:` in
https://github.com/inspect-js/is-core-module/commit/73412230a769f6e81c05eea50b6520cebf54ed2f
require `is-core-module@^2.3.0`.

Note: `is-core-module` checks whether the module is supported in a given
Node.js version (if not provided, the current version is checked).
Although checking whether a name is a core module in all Node.js
versions supported by the package would be ideal, it will require
significantly more design and implementation work and is therefore not
included in this commit.

Fixes: #275

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>